confusion reigns on high
I guess I'm a little confused about what's going on with Hackage.
Previously I assumed that the latest release was 0.2.2 and all these
changes were getting rolled int an 0.2.3 release. But I guess I had
released an 0.2.3 already without ever making an 0.2.3 git tag? And now
the git history is bonkers, with the commit that adds 0.2.3's alter
implementation coming after commits that weren't in the 0.2.3 release,
and I don't know how that happened.

Anyway. This commit bumps the version to 0.2.4 and attempts to fix up
the documentation and other stuff properly. There's probably still
errors in there but what can you do.